Utilize `usethis::proj_set()` instead of `usethis::local_project()`
Second time writing this commit message since it got lost the first time =( I'm
now writing this commit on a text editor window just to be safe.

I see that `usethis::local_project()` runs `usethis::proj_set()` twice: first
for the new `path`, then inside a `withr::defer()` call that restores the
original path. I don't fully understand `withr::defer()`, but it seems to be
triggered earlier that I would like it to. That's why I'm seeing on the
README, some examples, and the vignette after
https://github.com/r-lib/usethis/issues/1108#issuecomment-626262048
paths that are not the one I expected (the one equivalent to
`file.path(tempdir(), "biocthisexample")`).

One option would be to rely on `usethis::with_project()` to get a similar
behavior to what I had with `withr::with_dir()`, however like Jenny Bryan
point out, this makes the code confusing for a new user to understand. That's
why it seems to me that `usethis::proj_set()` is the better option to use
in the examples, README and vignette code.

I'm also now using a second example package name in the vignette to avoid
clashing with the examples. The code doesn't overwrite the files by default,
so if you run the code with the same example twice you don't get to see the
output messages the second time.