Improve workflow steps after working through them
Today with @lahuuki and @joshstolz we found several typos and steps that were not described in the dev template files. Also, the vignette did not include the installation steps for the development version of `biocthis` (it assumed that the package was already available from Bioconductor which is not the case).

Related to:

https://github.com/lahuuki/testtest/commit/ad865eba7bd3db3264d1811d69efbcf6aee11905
https://github.com/lahuuki/testtest/commit/db5b477cf779a1e03476057907cd2191b744c785
https://github.com/lahuuki/testtest/commit/f777e89876e4ab2687063426a40ade504f6d0ce7